+++ src/graph/executor/algo/AllPathsExecutor.h
@@ -5,20 +5,15 @@
 #ifndef GRAPH_EXECUTOR_ALGO_ALLPATHSEXECUTOR_H_
 #define GRAPH_EXECUTOR_ALGO_ALLPATHSEXECUTOR_H_
 
-#include "graph/executor/Executor.h"
-
-using nebula::storage::StorageRpcResponse;
-using nebula::storage::cpp2::GetNeighborsResponse;
-using RpcResponse = StorageRpcResponse<GetNeighborsResponse>;
-using PropRpcResponse = StorageRpcResponse<nebula::storage::cpp2::GetPropResponse>;
-using nebula::storage::StorageClient;
+#include "graph/executor/PathBaseExecutor.h"
 
 namespace nebula {
 namespace graph {
 class AllPaths;
-class AllPathsExecutor final : public Executor {
+class AllPathsExecutor final : public PathBaseExecutor {
  public:
-  AllPathsExecutor(const PlanNode* node, QueryContext* qctx) : Executor("AllPaths", node, qctx) {}
+  AllPathsExecutor(const PlanNode* node, QueryContext* qctx)
+      : PathBaseExecutor("AllPaths", node, qctx) {}
 
   folly::Future<Status> execute() override;
 
@@ -47,80 +42,6 @@ class AllPathsExecutor final : public Executor {
 
   folly::Future<Status> buildResult();
 
-  bool hasSameEdge(const std::vector<Value>& edgeList, const Edge& edge);
-
-  std::vector<Value> handlePropResp(PropRpcResponse&& resps);
-
-  folly::Future<std::vector<Value>> getProps(const std::vector<Value>& vids);
-
-  Status handleErrorCode(nebula::cpp2::ErrorCode code, PartitionID partId) const;
-
-  struct VertexHash {
-    std::size_t operator()(const Value& v) const {
-      switch (v.type()) {
-        case Value::Type::VERTEX: {
-          auto& vid = v.getVertex().vid;
-          if (vid.type() == Value::Type::STRING) {
-            return std::hash<std::string>()(vid.getStr());
-          } else {
-            return vid.getInt();
-          }
-        }
-        case Value::Type::STRING: {
-          return std::hash<std::string>()(v.getStr());
-        }
-        case Value::Type::INT: {
-          return v.getInt();
-        }
-        default: {
-          return v.hash();
-        }
-      }
-    }
-  };
-
-  struct VertexEqual {
-    bool operator()(const Value& lhs, const Value& rhs) const {
-      if (lhs.type() == rhs.type()) {
-        if (lhs.isVertex()) {
-          return lhs.getVertex().vid == rhs.getVertex().vid;
-        }
-        return lhs == rhs;
-      }
-      if (lhs.type() == Value::Type::VERTEX) {
-        return lhs.getVertex().vid == rhs;
-      }
-      if (rhs.type() == Value::Type::VERTEX) {
-        return lhs == rhs.getVertex().vid;
-      }
-      return lhs == rhs;
-    }
-  };
-
-  template <typename Resp>
-  StatusOr<Result::State> handleCompleteness(const storage::StorageRpcResponse<Resp>& rpcResp,
-                                             bool isPartialSuccessAccepted) const {
-    auto completeness = rpcResp.completeness();
-    if (completeness != 100) {
-      const auto& failedCodes = rpcResp.failedParts();
-      for (auto it = failedCodes.begin(); it != failedCodes.end(); it++) {
-        LOG(ERROR) << " failed, error " << apache::thrift::util::enumNameSafe(it->second)
-                   << ", part " << it->first;
-      }
-      // cannot execute at all, or partial success is not accepted
-      if (completeness == 0 || !isPartialSuccessAccepted) {
-        if (failedCodes.size() > 0) {
-          return handleErrorCode(failedCodes.begin()->second, failedCodes.begin()->first);
-        }
-        return Status::Error("Request to storage failed, without failedCodes.");
-      }
-      // partial success is accepted
-      qctx_->setPartialSuccess();
-      return Result::State::kPartialSuccess;
-    }
-    return Result::State::kSuccess;
-  }
-
  private:
   const AllPaths* pathNode_{nullptr};
   bool withProp_{false};
@@ -140,7 +61,6 @@ class AllPathsExecutor final : public Executor {
 
   DataSet result_;
   std::vector<Value> emptyPropVids_;
-  // folly::SpinLock statsLock_;
 };
 }  // namespace graph
 }  // namespace nebula
